+++ CHANGELOG.md
@@ -21,7 +21,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 -   Fixed measure tools
 -   State filter works
--
+-   Open Street Maps checked on when map loads
+-   Filter query returns no results when the only sites matching query are invalid
 
 ## [v0.4.0](https://github.com/USGS-WiM/stnweb2/releases/tag/v0.4.0) - 2020-01-12